Coffee list on Home page, modal dialog

+++ src/main/java/coffeeShop/controllers/ManufacturerController.java
@@ -1,29 +1,33 @@
 package coffeeShop.controllers;
 
 import coffeeShop.services.ManufacturerService;
+import coffeeShop.services.ProductService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
 
 @Controller
 public class ManufacturerController {
     private ManufacturerService manufacturerService;
-    public ManufacturerController(ManufacturerService manufacturerService) {
+    private ProductService productService;
+    public ManufacturerController(ManufacturerService manufacturerService,
+                                  ProductService productService) {
         this.manufacturerService = manufacturerService;
+        this.productService = productService;
     }
 
     @GetMapping("/manufacturers")
     public String getManufacturers(Model model){
         model.addAttribute("manufacturers", manufacturerService.list());
+        model.addAttribute("errorMessage", manufacturerService.getErrorMessage());
+
         return "admin/manufacturers";
     }
 
     @PostMapping("/manufacturers")
-    public String createSupplier(@RequestParam String name,
+    public String createSupplier(Model model, @RequestParam String name,
                                  @RequestParam String url) {
+        model.addAttribute("errorMessage", manufacturerService.getErrorMessage());
         if (!name.isEmpty() && !url.isEmpty()) {
             this.manufacturerService.add(name, url);
             return "redirect:/manufacturers";
@@ -41,7 +45,6 @@ public class ManufacturerController {
     public String updateManufacturer(@PathVariable(value = "manufacturerId") long id,
                                  @RequestParam String name,
                                  @RequestParam String url){
-//                                 @RequestParam ArrayList<Product>products){
         this.manufacturerService.change(name, url, id);
 
         return "redirect:/manufacturer-edit/{manufacturerId}";
@@ -49,8 +52,18 @@ public class ManufacturerController {
 
     @PostMapping("/manufacturers/{manufacturerId}")
     public String delete(@PathVariable(value = "manufacturerId") long manufacturerId) {
-        this.manufacturerService.remove(manufacturerId);
+        if(productService.getProductsByManufacturer(manufacturerId).isEmpty()){
+            this.manufacturerService.remove(manufacturerId);
+        }
+        else {
+            manufacturerService.setErrorMessage("Ei ole mahdollista poistaa valmistajaa, koska tietokannassa on vielä tämän valmistajan tuotteita.");
+        }
         return "redirect:/manufacturers";
     }
 
+    @GetMapping("/clearErrorMessage")
+    @ResponseBody
+    public void clearErrorMessage() {
+        manufacturerService.clearErrorMessage();
+    }
 }
